# api/trustgraph.openapi.yaml
openapi: 3.1.0
info:
  title: TrustGraph API
  version: 0.2.0
  description: |
    Trust, safety, and verification service for ConnectionSphere.

    Planes:
      - Plane A: first-party signals and trust assessment.
      - Plane B: consented verification (LinkedIn, government ID, liveness,
        image). Every Plane B read/write requires an active consent.
      - Plane C: investigation tools, restricted to authenticated investigators.

    Auth:
      - Plane A assessment endpoints are called service-to-service.
      - Plane B endpoints are subject-scoped and reached by the app on behalf
        of the subject; consent is enforced server-side per operation.
      - `/v1/admin/*` and `/v1/metrics/*` require `Authorization: Bearer
        $ADMIN_TOKEN`.
      - `/v1/investigations/*` require `Authorization: Bearer
        $INVESTIGATOR_TOKEN` (or `$ADMIN_TOKEN` for supervisor-level access),
        plus an optional `X-Investigator-Actor` header naming the human.

    When a vendor integration is not configured, its endpoint returns
    `503 not_configured`. It never returns a passing result it did not obtain.
servers:
  - url: http://localhost:8081
    description: Local development (via ConnectionSphere Docker network)
  - url: https://trustgraph-api.onrender.com
    description: Render production
...
paths:
...
  /v1/investigations/break-glass:
    post:
      operationId: requestBreakGlass
      summary: Request time-boxed emergency elevated access
      tags: [Investigations (Plane C)]
      description: |
        A justification of at least 20 characters is required. The grant is
        time-boxed (default 60 minutes, capped at 4 hours), recorded in
        `investigation_break_glass`, and flagged for review in the access log.
      security:
        - InvestigatorBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [justification]
              properties:
                justification:
                  type: string
                  minLength: 20
                subjectId:
                  type: string
                  format: uuid
                caseId:
                  type: string
                  format: uuid
      responses:
        "201":
          description: Grant created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BreakGlassGrant"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "503":
          $ref: "#/components/responses/NotConfigured"
components:
...
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Missing or malformed bearer token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    NotConfigured:
      description: The backing vendor integration is not configured on this deployment
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  securitySchemes:
    ServiceAccountHeader:
      type: apiKey
      in: header
      name: X-Service-Account
      description: ConnectionSphere service account identifier
    AdminBearer:
      type: http
      scheme: bearer
      description: "Set via ADMIN_TOKEN. Grants supervisor-level Plane C access."
    InvestigatorBearer:
      type: http
      scheme: bearer
      description: "Set via INVESTIGATOR_TOKEN. Optional X-Investigator-Actor header names the human."
  schemas:
...
    BreakGlassGrant:
      type: object
      properties:
        grantId:
          type: string
          format: uuid
        actor:
          type: string
        actorRole:
          type: string
        subjectId:
          type: string
          format: uuid
        caseId:
          type: string
          format: uuid
        justification:
          type: string
        grantedAt:
          type: string
          format: date-time
        expiresAt:
          type: string
          format: date-time
        revokedAt:
          type: string
          format: date-time
        reviewedBy:
          type: string
        reviewedAt:
          type: string
          format: date-time
        reviewOutcome:
          type: string
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: bad_request
        message:
          type: string